backend/requirements.txt
Django>=4.2.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
channels>=4.0.0
channels-redis>=4.1.0
celery>=5.3.4
redis>=5.0.1
psycopg2-binary>=2.9.7
djangorestframework-simplejwt>=5.3.0
django-environ>=0.10.0
mnemonic>=0.20
bcrypt>=4.0.1
sendgrid>=6.10.0
drf-yasg>=1.21.7
django-filter>=23.3
whitenoise>=6.6.0
gunicorn>=21.2.0
Pillow>=10.1.0
django-celery-results>=2.5.1